Moved static picture data to pictures api controller (picture.controller.js)
