+++ server/api/v1/picture/picture.controller.js
@@ -1,12 +1,446 @@
 'use strict';
 
-//var mockData = require("./mockData.js");
-
 // Get list of pictures
 exports.index = index;
 
 function index(req, res) {
 
-  //var payload = mockData.loadMockPictures();
-  res.json({"hello": "world!!!"});
+  var data = {
+    "meta": {
+      "paging": {
+        "limit": 52,
+        "offset": 0,
+        "total": 208,
+        "links": {
+          "next": '/api/v1/pictures?limit=52&offset=52',
+          "first": '/api/v1/pictures?limit=52&offset=0',
+          "last": '/api/v1/pictures?limit=52&offset=156',
+          "current": '/api/v1/pictures?limit=52&offset=0'
+        }
+      },
+      "result": {
+        "status": 'ok'
+      }
+    },
+    "pictures": [
+      {
+        "id": '16414316734',
+        "title": '5D3_7127 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8772/16414316734_25428bc7d7.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850528669',
+        "title": '5D3_7114 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8808/16850528669_b01e0b57a8.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16829340237',
+        "title": '5D3_7111 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8761/16829340237_b41b271931.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035888361',
+        "title": '5D3_7108 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8798/17035888361_a8be945b6a.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850525149',
+        "title": '5D3_7104 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7698/16850525149_cb47c87930.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17036740385',
+        "title": '5D3_7089 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7727/17036740385_05495a02ef.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035294962',
+        "title": '5D3_7052 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8817/17035294962_571d9e3145.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850521969',
+        "title": '5D3_7049 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7671/16850521969_68dd471c96.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414308144',
+        "title": '5D3_7031 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7665/16414308144_c6ddefb45e.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17036736185',
+        "title": '5D3_7030 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8723/17036736185_b9d3f01d24.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850518769',
+        "title": '5D3_7024 - Version 3',
+        "url": 'https://farm9.staticflickr.com/8816/16850518769_184474d25a.jpg',
+        "width": 500,
+        "height": 281,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414303644',
+        "title": '5D3_7018 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8726/16414303644_1ae24c5f90.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16848943308',
+        "title": '5D3_7007 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7614/16848943308_4a8f325bbd.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16829327347',
+        "title": '5D3_6986 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7613/16829327347_7571572395.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17010733616',
+        "title": '5D3_6892 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7654/17010733616_76306a1f25.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16848939898',
+        "title": '5D3_6847 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8714/16848939898_6b5e3d783c.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16848938898',
+        "title": '5D3_6795 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8784/16848938898_de2a69b1b7.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035872131',
+        "title": '5D3_6792 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7640/17035872131_5cecf7227f.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850508359',
+        "title": '5D3_6791 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8771/16850508359_6815ac4bb4.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416573453',
+        "title": '5D3_6787 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7590/16416573453_afba52330d.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850506249',
+        "title": '5D3_6738 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8716/16850506249_9afdc1b207.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17036721815',
+        "title": '5D3_6731 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7719/17036721815_65a12aca22.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850503119',
+        "title": '5D3_6729 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7705/16850503119_7da6bb1544.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850502169',
+        "title": '5D3_6728 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7655/16850502169_533acc4873.jpg',
+        "width": 500,
+        "height": 281,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416567503',
+        "title": '5D3_6725 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8820/16416567503_cb83bc37bd.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035862621',
+        "title": '5D3_6720 - Version 3',
+        "url": 'https://farm8.staticflickr.com/7597/17035862621_03f99a27d5.jpg',
+        "width": 500,
+        "height": 281,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16849186280',
+        "title": '5D3_6709 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8772/16849186280_938af1d09a.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16829311557',
+        "title": '5D3_6708 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7722/16829311557_843e42e46c.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850496489',
+        "title": '5D3_6702 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8745/16850496489_c41b26ed78.jpg',
+        "width": 500,
+        "height": 281,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17010716826',
+        "title": '5D3_6697 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7621/17010716826_d124b4a21a.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414281094',
+        "title": '5D3_6693 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8785/16414281094_bc2d94fe54.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17010714396',
+        "title": '5D3_6691 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8801/17010714396_a5dd068c2a.jpg',
+        "width": 500,
+        "height": 281,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416556763',
+        "title": '5D3_6685 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8786/16416556763_ec99d14753.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16829303367',
+        "title": '5D3_6684 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8756/16829303367_319b429db0.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035850591',
+        "title": '5D3_6682 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8748/17035850591_9d1f83f6d7.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16850487419',
+        "title": '5D3_6677 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8754/16850487419_024505ce11.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414274214',
+        "title": '5D3_6675 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7592/16414274214_f6a8c3960c.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035847211',
+        "title": '5D3_6674 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7722/17035847211_c59660ffa7.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17036700375',
+        "title": '5D3_6671 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8695/17036700375_f7af003f37.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16848911608',
+        "title": '5D3_6664 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7599/16848911608_ed0dc9c7f4.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035254732',
+        "title": '5D3_6660 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8728/17035254732_a6fc09a9c6.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16848908918',
+        "title": '5D3_6656 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7664/16848908918_776bac9243.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16829291047',
+        "title": '5D3_6653 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8801/16829291047_8b682cc2e9.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416542503',
+        "title": '5D3_6651 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8713/16416542503_6c4349f909.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416541213',
+        "title": '5D3_6649 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8687/16416541213_9fb7acac78.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035246932',
+        "title": '5D3_6646',
+        "url": 'https://farm9.staticflickr.com/8738/17035246932_61946cf539.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16416538743',
+        "title": '5D3_6644',
+        "url": 'https://farm9.staticflickr.com/8731/16416538743_f953fd429b.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414259124',
+        "title": '5D3_6641 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7625/16414259124_e871f74126.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414257854',
+        "title": '5D3_6637 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8685/16414257854_61568370b2.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035830891',
+        "title": '5D3_6633',
+        "url": 'https://farm9.staticflickr.com/8809/17035830891_f5a470b41f.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '16414255344',
+        "title": '5D3_6630 - Version 2',
+        "url": 'https://farm8.staticflickr.com/7662/16414255344_b56ec6e4e7.jpg',
+        "width": 333,
+        "height": 500,
+        "tags": 'wedding bentley'
+      },
+      {
+        "id": '17035239782',
+        "title": '5D3_6626 - Version 2',
+        "url": 'https://farm9.staticflickr.com/8730/17035239782_2139255364.jpg',
+        "width": 500,
+        "height": 333,
+        "tags": 'wedding bentley'
+      }
+    ]
+  };
+
+  res.json(data);
 }
